Finished Phase 5: Fixed the page division of the address column. 	modified:   include/vdb/common/types.h 	modified:   include/vdb/storage/address_column.h 	modified:   include/vdb/storage/data_file_writer.h 	modified:   src/storage/address_column.cpp 	modified:   src/storage/data_file_writer.cpp 	modified:   tests/storage/address_column_test.cpp 	modified:   tests/storage/cluster_store_test.cpp 	modified:   tests/storage/data_file_test.cpp 	modified:   tests/storage/segment_test.cpp

+++ include/vdb/common/types.h
@@ -470,8 +470,8 @@ struct RaBitQConfig {
 /// Default block size for I/O alignment (4KB)
 constexpr size_t kDefaultBlockSize = 4096;
 
-/// Default page size for mmap
-constexpr size_t kDefaultPageSize = 4096;
+/// Default page size for page-aligned record storage (4KB)
+constexpr uint32_t kDefaultPageSize = 4096;
 
 /// Cache line size for SIMD alignment
 constexpr size_t kCacheLineSize = 64;
